pyproject.toml
[tool.black]
line-length = 100
target-version = ['py36']
quiet = true
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.venv
  | _build
  | build
  | dist
  | [a-z_]+/migrations
)/
'''

[tool.bandit]
exclude_dirs = ["*/test_*.py", "*/tests.py", "tests", "./venv/", "./env/", "./node_modules/", "./cacheback/", "./.env", "./.venv"]
skips = ["B101", "B311"]

[tool.coverage.run]
branch = true
source = ["backend"]
omit = ["**/virtualenvs/*", "**/migrations/*", "**/settings/*", "**/tests/*"]

[tool.isort]
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_django = "django"
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
lines_after_imports = 2
atomic = true
combine_star = true
skip = ["venv", "env", "node_modules", "migrations", ".env", ".venv"]

[tool.poetry]
name = "{{project_name}}"
version = "0.1.0"
description = "Django 3, React, Bootstrap 5 with Python 3 and webpack project boilerplate"
authors = ["Vinta Software <contact@vinta.com.br>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = ">=3,<4"
celery = {extras = ["redis"], version = "^5.3.0"}
celery-redbeat = "^2.1.1"
django-model-utils = "^4.3.1"
django-webpack-loader = "^2.0.1"
django-js-reverse = "^0.9.1"
django-import-export = "^3.2.0"
djangorestframework = "^3.14.0"
django-debreach = "^2.1.0"
python-decouple = "^3.8"
psycopg2 = "^2.9.6"
brotlipy = "^0.7.0"
django-log-request-id = "^2.1.0"
dj-database-url = "^2.0.0"
gunicorn = "^20.1.0"
whitenoise = "^6.4.0"
psutil = "^5.9.5"
ipython = "^8.14.0"
sentry-sdk = "^1.25.1"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.5"
black = "^23.3.0"
coverage = "^7.2.7"
isort = "^5.12.0"
model-bakery = "^1.12.0"
pre-commit = "^3.3.3"
pylint-django = "^2.5.3"
prospector = {extras = ["with-vulture"], version = "^1.10.2"}

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
